Update the project to include an authorization handler
